multiple changes to CMMR

+++ Backend/controllers/documentsController.js
@@ -1,5 +1,21 @@
 const Document = require("../models/document");
 
+exports.uploadDocument = async (req, res) => {
+  try {
+    // Create a new document record in the database
+    const document = new Document({
+      title: req.body.title,
+      abstract: req.body.abstract,
+      user: req.body.userId,
+      filePath: req.file.path, // Store the file path in the database
+    });
+    await document.save();
+    res.status(201).json(document);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
 exports.getAllDocuments = async (req, res) => {
   try {
     const documents = await Document.find();
@@ -25,6 +41,10 @@ exports.getDocument = async (req, res) => {
 
 exports.createDocument = async (req, res) => {
   try {
+    const user = req.user;
+    if (user.role !== "admin") {
+      return res.status(403).json({ message: "Unauthorized" });
+    }
     const document = new Document(req.body);
     await document.save();
     res.status(201).json(document);
@@ -36,6 +56,10 @@ exports.createDocument = async (req, res) => {
 
 exports.updateDocument = async (req, res) => {
   try {
+    const user = req.user;
+    if (user.role !== "admin") {
+      return res.status(403).json({ message: "Unauthorized" });
+    }
     const document = await Document.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
@@ -51,6 +75,10 @@ exports.updateDocument = async (req, res) => {
 
 exports.deleteDocument = async (req, res) => {
   try {
+    const user = req.user;
+    if (user.role !== "admin") {
+      return res.status(403).json({ message: "Unauthorized" });
+    }
     const document = await Document.findByIdAndDelete(req.params.id);
     if (!document) {
       return res.status(404).json({ message: "Document not found" });
